Refactor resolver to support AXFR queries
Pass a pointer to the record to be resolved so we can update it with the
name and type of the RRSIG expiring first.  This won't affect existing
behaviour.  Even when multiple RRSIGs exist for a given name/type, e.g.
during key rollovers, they'll all have the same name/type.

Add a test to ensure the existing behaviour is preserved.

Expand the definition of "resolves" to include authoritative responses.
Authoritative resolvers set the AA bit.  They don't validate signatures
and won't set the AD bit.

While here, remove the unnecessary test for Checking Disabled when
determining if a record resolves.  AD implies !CD.

Add tests for AXFR queries and authoritative responses.

These changes do not (or should not) change any existing behaviour for
realistic use cases (i.e. queries to validating resolvers).